* master:
  Feature/deseralization (#1)
  adding remove function tests
  remove function
  adding tests; let script compile all the files
  function comments
  adding Length, Containt and empty Remove functions

# Conflicts:
#	list.go
#	list_test.go
#	main.go

+++ node.go
@@ -2,6 +2,6 @@ package main
 
 // Node is a single element in a list
 type Node struct {
-	Value uint
+	Value uint16
 	Next  *Node
 }
